Refactor to use Express/Apollo instead of Meteor
This is a really huge commit that replaces Meteor with a combination of
Express for the server part, GraphQL/Apollo for the queries and Webpack
for building the client side of the application. Right now the RPC calls
are done using REST queries. The directory structure has also been
reorganized. The build works but there are still many bugs.